+++ src/OmniSharp/Api/Highlighting/OmnisharpController.Highlighting.cs
@@ -25,7 +25,7 @@ namespace OmniSharp
             var root = await tree.GetRootAsync();
 
             var nodes = new List<SyntaxNode>();
-            if (request.Lines.Length == 0)
+            if (request.Lines == null || request.Lines.Length == 0)
             {
                 nodes.Add(root);
             }
@@ -36,12 +36,15 @@ namespace OmniSharp
                 {
                     var lineSpan = text.Lines[line].Span;
                     var node = root;
-                    var start = node.ChildThatContainsPosition(lineSpan.Start);
-                    var end = node.ChildThatContainsPosition(lineSpan.End);
-                    while (start.IsNode && start == end)
+                    var next = node;
+                    SyntaxNodeOrToken start, end;
+                    do
                     {
-                        node = start.AsNode();
-                    }
+                        node = next;
+                        start = node.ChildThatContainsPosition(lineSpan.Start);
+                        end = node.ChildThatContainsPosition(lineSpan.End);
+                        next = start.AsNode();
+                    } while (start.IsNode && start == end);
 
                     nodes.Add(node);
                 }
@@ -53,187 +56,13 @@ namespace OmniSharp
                 walker.Visit(node);
             }
 
-            return walker.Regions;
-        }
-
-        class HighlightSyntaxWalker : CSharpSyntaxWalker
-        {
-            readonly SemanticModel _model;
-            readonly List<HighlightResponse> _regions;
-
-            // TODO: Fix overlap
-            public IImmutableList<HighlightResponse> Regions
+            var regions = walker.Regions;
+            if (request.Lines != null && request.Lines.Length != 0)
             {
-                get
-                {
-                    return _regions
-                        .GroupBy(r => new { r.Line, r.Start, r.End })
-                        .Select(g => new HighlightResponse
-                        {
-                            Line = g.Key.Line,
-                            Start = g.Key.Start,
-                            End = g.Key.End,
-                            Kind = string.Join(" ", g.Select(r => r.Kind).Distinct())
-                        })
-                        .ToImmutableList();
-                }
+                return regions.Where(r => request.Lines.Contains(r.Line));
             }
 
-            public HighlightSyntaxWalker(SemanticModel model)
-                : base(SyntaxWalkerDepth.Trivia)
-            {
-                _model = model;
-            }
-
-            void Mark(SyntaxToken token, string type = null)
-            {
-                var location = token.GetLocation().GetLineSpan();
-                var startLine = location.StartLinePosition.Line;
-                var endLine = location.EndLinePosition.Line;
-
-                for (var i = startLine; i <= endLine; i++)
-                {
-                    var start = i == startLine ? location.StartLinePosition.Character : 0;
-                    var end = i == endLine ? location.EndLinePosition.Character : int.MaxValue;
-
-                    _regions.Add(new HighlightResponse
-                    {
-                        Line = i,
-                        Start = start,
-                        End = end,
-                        Kind = type
-                    });
-                }
-            }
-
-            void Mark(SyntaxTrivia trivia, string type = null)
-            {
-                var location = trivia.GetLocation().GetLineSpan();
-                var startLine = location.StartLinePosition.Line;
-                var endLine = location.EndLinePosition.Line;
-
-                for (var i = startLine; i <= endLine; i++)
-                {
-                    var start = i == startLine ? location.StartLinePosition.Character : 0;
-                    var end = i == endLine ? location.EndLinePosition.Character : int.MaxValue;
-
-                    _regions.Add(new HighlightResponse
-                    {
-                        Line = i,
-                        Start = start,
-                        End = end,
-                        Kind = type
-                    });
-                }
-            }
-
-            ISymbol GetTokenSymbol(SyntaxToken token)
-            {
-                var symbol = _model.GetDeclaredSymbol(token.Parent);
-                if (symbol == null)
-                {
-                    // The token isnt part of a declaration node, so try to get symbol info.
-                    symbol = _model.GetSymbolInfo(token.Parent).Symbol;
-                    if (symbol == null)
-                    {
-                        // we couldnt find symbol information for the node, so we will look at all symbols in scope by name.
-                        var namedSymbols = _model.LookupSymbols(token.SpanStart, null, token.ToString(), true);
-                        if (namedSymbols.Length == 1)
-                        {
-                            symbol = namedSymbols[0];
-                        }
-                    }
-                }
-
-                return symbol;
-            }
-
-            void MarkIdentifier(SyntaxToken token)
-            {
-                var symbol = GetTokenSymbol(token);
-                if (symbol == null)
-                {
-                    Mark(token, "identifier");
-                }
-                else
-                {
-                    VisitSymbol(token, symbol);
-                }
-            }
-
-            void VisitSymbol(SyntaxToken token, ISymbol symbol)
-            {
-                var parts = symbol.ToDisplayParts(new SymbolDisplayFormat());
-                var part = parts.SingleOrDefault(p => p.Symbol == symbol);
-                if (part.Symbol != null)
-                {
-                    Mark(token, part.Kind.ToString().ToLowerInvariant());
-                }
-            }
-
-            public override void VisitToken(SyntaxToken token)
-            {
-                if (token.IsKeyword() || token.IsContextualKeyword())
-                {
-                    Mark(token, "keyword");
-                    var symbol = GetTokenSymbol(token);
-                    if (symbol != null)
-                    {
-                        VisitSymbol(token, symbol);
-                    }
-                }
-
-                var kind = token.Kind();
-                switch (kind)
-                {
-                    case SyntaxKind.IdentifierToken:
-                        MarkIdentifier(token); break;
-
-                    case SyntaxKind.StringLiteralToken:
-                        Mark(token, "string"); break;
-
-                    case SyntaxKind.NumericLiteralToken:
-                        Mark(token, "number"); break;
-
-                    case SyntaxKind.CharacterLiteralToken:
-                        Mark(token, "char"); break;
-                }
-
-                base.VisitToken(token);
-            }
-
-            public override void VisitTrivia(SyntaxTrivia trivia)
-            {
-                if (trivia.HasStructure)
-                    Visit(trivia.GetStructure());
-
-                switch (trivia.Kind())
-                {
-                    case SyntaxKind.SingleLineDocumentationCommentTrivia:
-                    case SyntaxKind.MultiLineDocumentationCommentTrivia:
-                    case SyntaxKind.DocumentationCommentExteriorTrivia:
-                        Mark(trivia, "doc");
-
-                        goto case SyntaxKind.SingleLineCommentTrivia;
-
-                    case SyntaxKind.MultiLineCommentTrivia:
-                    case SyntaxKind.SingleLineCommentTrivia:
-                        Mark(trivia, "comment"); break;
-
-                    case SyntaxKind.EndOfLineTrivia:
-                    case SyntaxKind.WhitespaceTrivia:
-                        Mark(trivia, "whitespace"); break;
-
-                    case SyntaxKind.RegionDirectiveTrivia:
-                    case SyntaxKind.EndRegionDirectiveTrivia:
-                        Mark(trivia, "region"); break;
-
-                    case SyntaxKind.DisabledTextTrivia:
-                        Mark(trivia, "disabled-text"); break;
-                }
-
-                base.VisitTrivia(trivia);
-            }
+            return regions;
         }
     }
 }
